Rewrite CMake configuration to use target based dependency resolution.
The overall goal is modernization and simplification of CMake
configuration to make it easier to maintain. The dependency resolution
is now based on targets and was tested in standalone linux build
with resolution using pkg-config and as part of the engine build.

A lot of features were dropped, as they don't seem to be useful anymore,
and when they will be, they can be re-added in a more modern way.
Removed:
- direct dependency resolution for dependencies of dependencies
- a bunch of optional compilation flags to simplify, those can be
  re-added back when proven useful and needed again
- windows icon setting: it was broken anyway, only a nice to have
- building both shared and static pr-downloader binary in every build
- installation of pr-downloader header files and libraries, keeping
  only the executable.
- setting of different compiler flags

The current shape is not perfect, and will be iterated on, one
of the main problems being compatibility with current engine build.