Support for C++ 20 compilation
C++ 20 introduces a std::u8string type which is not implicitly coercable
to a std::string. Calls to std::filesystem::path::u8string() are thus no
longer compatible with the std::string type.

This change forces an explicit cast from std::u8string to std::string in
C++ 20 compilations. In C++ 17 compilations, the cast is from
std::string to std::string (a noop).